requirements.txt
ChatterBot==0.8.4
chatterbot-corpus==1.1.1
dj-database-url==0.4.2
Django==2.0.2
django-appconf==1.0.2
django-filter==1.1.0
django-imagekit==4.0.2
djangorestframework==3.7.7
future==0.16.0
gcloud==0.17.0
googleapis-common-protos==1.5.3
gunicorn==19.7.1
httplib2==0.10.3
jws==0.1.3
Markdown==2.6.11
mathparse==0.1.1
nltk==3.2.5
oauth2client==3.0.0
oauthlib==2.0.6
pilkit==2.0
Pillow==5.0.0
protobuf==3.5.1
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycryptodome==3.4.3
pymongo==3.6.0
Pyrebase==3.0.27
python-dateutil==2.6.1
python-jwt==2.0.1
python-twitter==3.3
pytz==2017.3
PyYAML==3.12
requests==2.11.1
requests-oauthlib==0.8.0
requests-toolbelt==0.7.0
rsa==3.4.2
six==1.11.0
SQLAlchemy==1.2.2
